requirements.txt
requests==2.32.3
python-dotenv==1.0.1
Flask==3.1.0
ordereddict==1.1
simplejson==3.19.3
unicode==2.9
pymongo==4.10.1